requirements.txt
python-telegram-bot==21.9
supabase==2.10.0
python-dotenv==1.0.0

